refactor: move push sign and upload key back to store
